Changes SetGravity to a Variable-Argument Spell
Changes OpSetGravity such that, when cast on the self, applies an expensive two-argument version which applies permanently (or at least until death or overridden); when cast on other entities, executes a cheaper three-argument version which applies only for the duration specified by the third argument.
Various support changes to allow that to work.